NAVCal

Documentation for NAVCal.

NAVCal.KelseyFerguson

NAVCal.MillBrook

NAVCal.RiverRoad

NAVCal.WellsRiver

NAVCal.annum

NAVCal.varve_year

NAVCal.KelseyFerguson — Method

KelseyFerguson()

Varve thickness and NAVC calibration data for the Kelsey Ferguson Brickyard site in South Windsor, CT.

key	description

:sequence	NAVC sequence name

:a	NAVC varve year

:z	varve (couplet) thickness (cm)

:summer	summer layer thickness (cm)

:winter	winter layer thickness (cm)

:transect	Outcrop transit number

:notes	Associated notes

NAVCal.MillBrook — Method

MillBrook()

Varve thickness and NAVC calibration data from the Mill Brook site near Putney, Vt.

key	description

:sequence	NAVC sequence name

:a	NAVC varve year

:z	varve thickness (cm)

NAVCal.RiverRoad — Method

RiverRoad()

Varve thickness and NAVC calibration data from the River Road site, as measured by Levy (1998).

key	description

:sequence	NAVC sequence name

:a	NAVC varve year

:z	varve thickness (cm)

:notes	Associated notes

NAVCal.WellsRiver — Function

WellsRiver(n)

Varve thickness and NAVC calibration data from the Wells River Valley, near Wells River, VT.

You must provide an Int 1-5, corresponding to the files below:

ID	description	NAVC age range

1	Sequence 1A	6772-7022

2	Sequence 1B	7024-7046

3	Sequence 1C	7049-7138

4	Sequence 1D	7157-7289

5	Sequence 2A	6772-7022

key	description

:sequence	NAVC sequence name

:a	NAVC varve year

:z	varve (couplet) thickness (cm)

:summer	summer layer thickness (cm)

:winter	winter layer thickness (cm)

:core	Core identifier

:meas	? (unclear data heading "Meas")

:notes	Associated notes

NAVCal.annum — Method

annum(x)

Convert NAVC varve year to age in annum.

NAVCal.varve_year — Method

varve_year(x)

Convert age in annum to NAVC varve year.